fix: fix change plantuml output directory
PlantUML doesn't let the user choose the name of the generated file,
only the directory. Also it capitalizes all words and change dashes to
spaces (maybe more chars conversions to spaces should be implemented)

TODO: look at PlantUML file name choice logic to re-implement it.
TODO: propose a PR to PlantUML to be able to change the file name.